+++ core/camel-core/src/test/java/org/apache/camel/builder/xml/DefaultNamespaceContextTest.java
@@ -21,6 +21,7 @@ import java.util.Iterator;
 import java.util.Map;
 
 import org.apache.camel.ContextTestSupport;
+import org.apache.camel.support.builder.xml.Namespaces;
 import org.junit.Test;
 
 public class DefaultNamespaceContextTest extends ContextTestSupport {
