+++ core/camel-core/src/test/java/org/apache/camel/builder/xml/XPathNamespaceResultTypeTest.java
@@ -17,6 +17,7 @@
 package org.apache.camel.builder.xml;
 
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.support.builder.xml.Namespaces;
 
 /**
  * XPath with namespace test
