+++ core/camel-core/src/test/java/org/apache/camel/builder/xml/XPathNamespaceTest.java
@@ -19,6 +19,7 @@ package org.apache.camel.builder.xml;
 import org.apache.camel.ContextTestSupport;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.support.builder.xml.Namespaces;
 import org.junit.Test;
 
 /**
