+++ core/camel-core/src/test/java/org/apache/camel/builder/xml/XPathNestedNamespaceTest.java
@@ -18,6 +18,7 @@ package org.apache.camel.builder.xml;
 
 import org.apache.camel.ContextTestSupport;
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.support.builder.xml.Namespaces;
 import org.junit.Test;
 
 public class XPathNestedNamespaceTest extends ContextTestSupport {
