+++ core/camel-core/src/test/java/org/apache/camel/language/tokenizer/XMLTokenizeWrapLanguageTest.java
@@ -18,7 +18,7 @@ package org.apache.camel.language.tokenizer;
 
 import org.apache.camel.ContextTestSupport;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.builder.xml.Namespaces;
+import org.apache.camel.support.builder.xml.Namespaces;
 import org.junit.Test;
 
 public class XMLTokenizeWrapLanguageTest extends ContextTestSupport {
