+++ core/camel-core/src/test/java/org/apache/camel/processor/SplitterWithXqureyTest.java
@@ -19,7 +19,7 @@ package org.apache.camel.processor;
 import org.apache.camel.ContextTestSupport;
 import org.apache.camel.Exchange;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.builder.xml.Namespaces;
+import org.apache.camel.support.builder.xml.Namespaces;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.Test;
