+++ core/camel-support/src/main/java/org/apache/camel/support/builder/xml/Namespaces.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.builder.xml;
+package org.apache.camel.support.builder.xml;
 
 import java.util.HashMap;
 import java.util.Map;
